format single job query
